Make separate components for Steps and StepControls for the Ledger account import flow
